Volgens mij kan ik wel committen want ik kan inloggen in mn eigen database

+++ TTNReceiver.Database/Measurements.sql
@@ -2,7 +2,7 @@
 (
 	[Id] INT NOT NULL PRIMARY KEY,
     [Name] varchar(50) NULL,
-    [DeviceId] INT NOT NULL, 
+    [DeviceId] SMALLINT NOT NULL, 
     [Timestamp] DATETIME2 NULL, 
     [Value] DECIMAL(18, 2) NULL, 
     CONSTRAINT [FK_Measurements_ToDevices] FOREIGN KEY ([DeviceId]) REFERENCES [Devices]([Id])
